split entry point event loop to main.c
